refactor(web): move diff page provider out of hooks
